update text summarization model
